+++ spec/lib/packages/virtualbox_guest_spec.rb
@@ -3,7 +3,7 @@ require 'spec_helper'
 describe 'packages/virtualbox_guest' do
   it 'adds the following commands' do
     eval_package
-    $conf.commands.map(&:info).join("\n").must_equal [
+    queued_commands.must_equal [
       "TASK   virtualbox_guest - Installs VirtualBox Guest Additions and fixes piix4_smbus error",
       "SUDO   echo deb http://download.virtualbox.org/virtualbox/debian precise contrib >> /etc/apt/sources.list",
       "SUDO   wget -q http://download.virtualbox.org/virtualbox/debian/oracle_vbox.asc -O - | apt-key add -",
@@ -16,7 +16,7 @@ describe 'packages/virtualbox_guest' do
   it 'autostart client when requested' do
     $conf.confgure_vbox_client = true
     eval_package
-    $conf.commands.map(&:info).join("\n").must_match /grep "VBoxClient-all &" .xinitrc || echo "VBoxClient-all &" >> .xinitrc/
+    queued_commands.must_match /grep "VBoxClient-all &" .xinitrc || echo "VBoxClient-all &" >> .xinitrc/
   end
 end